+++ src/converters/piral-mithril/src/converter.ts
@@ -6,13 +6,13 @@ import type { Component } from './types';
 export interface MithrilConverterOptions {
   /**
    * Defines the name of the root element.
-   * @default slot
+   * @default piral-slot
    */
   rootName?: string;
 }
 
 export function createConverter(config: MithrilConverterOptions = {}) {
-  const { rootName = 'slot' } = config;
+  const { rootName = 'piral-slot' } = config;
   const Extension = createExtension(rootName);
   const convert = <TProps extends BaseComponentProps>(
     component: Component<TProps>,