+++ src/converters/piral-react-15/src/converter.ts
@@ -6,13 +6,13 @@ import { mountReact15, unmountReact15 } from './mount';
 export interface React15ConverterOptions {
   /**
    * Defines the name of the root element.
-   * @default slot
+   * @default piral-slot
    */
   rootName?: string;
 }
 
 export function createConverter(config: React15ConverterOptions = {}) {
-  const { rootName = 'slot' } = config;
+  const { rootName = 'piral-slot' } = config;
   const Extension = createExtension(rootName);
   const convert = <TProps extends BaseComponentProps>(root: ComponentType<TProps>): ForeignComponent<TProps> => ({
     mount(el, props, ctx) {