+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Changed the default bundler to be `piral-cli-webpack5` (#469)
 - Changed webpack configs to use `oneOf` for assets (#451)
 - Changed the internal state container to `zustand`
+- Changed portal boundary to `piral-portal` using `display: contents`
+- Changed default extension slot to `piral-slot` using `display: contents`
 - Deprecated the usage of `renderInstance` (#465)
 - Deprecated the usage of `externals` for `pilets`
 - Removed support for IE11 (#467)